# 120x faster write_profile through buffered I/O & non-pretty printing