Serde serializes through tiny writes so essential to use BufWriter as File is unbuffered. 

This reduces writing down a 4.1 MB JSON profile report from 12840 ms to 168 ms. 

Also pretty printing is not important for this output nor for Chrome so disabled that as well and got down the total serialization time from 12840 ms to 107 ms. 

Not too shabby with a 120x speed of write_profile  for a simple 2 line change!