@repi 2018-01-06
Sweet. Would you be able to publish a new version on crates.io that one can depend on?