@glennw 2018-01-06
Sure - published 0.2.2 on crates.io.